+++ apkg/cache.py
@@ -4,23 +4,36 @@ apkg packaging file cache
 
 import json
 from pathlib import Path
+from typing import Dict, Iterable, List, Mapping, Union
 
+from apkg import __version__
 from apkg.log import getLogger
-from apkg.util.common import hash_file, hash_path
+from apkg.util.common import CacheableEntry, hash_file, hash_path
 
 
 log = getLogger(__name__)
 
 
-def file_checksum(*paths):
+CacheEntry = Union[str, int, bool, List['CacheEntry'], Dict[str, 'CacheEntry']]
+
+
+class FileMissingError(Exception):
+    pass
+
+
+class ChecksumError(Exception):
+    pass
+
+
+def file_checksum(*paths) -> str:
     return hash_file(*paths).hexdigest()[:20]
 
 
-def path_checksum(*paths):
+def path_checksum(*paths) -> str:
     return hash_path(*paths).hexdigest()[:20]
 
 
-def enabled_str(enabled):
+def enabled_str(enabled: bool) -> str:
     return 'ENABLED' if enabled else 'DISABLED'
 
 
@@ -31,18 +44,26 @@ class ProjectCache:
         self.cache = {}
         self.checksum = None
 
-    def save(self):
+    def save(self) -> None:
+        self.cache['__version__'] = __version__
         json.dump(self.cache, self.project.path.cache.open('w'))
 
-    def load(self):
+    def load(self) -> None:
         cache_path = self.project.path.cache
         if not cache_path.exists():
             log.verbose("cache not found: %s", cache_path)
             return
         log.verbose("loading cache: %s", cache_path)
-        self.cache = json.load(cache_path.open('r'))
 
-    def _ensure_load(self):
+        cache = json.load(cache_path.open('r'))
+        version = cache.get('__version__', None)
+        if version != __version__:
+            log.warning("ignoring cache from different apkg version: "
+                        "%s != %s", version, __version__)
+            return
+        self.cache = cache
+
+    def _ensure_load(self) -> None:
         """
         ensure cache is loaded on demand and only once
 
@@ -53,57 +74,54 @@ class ProjectCache:
         self.load()
         self.loaded = True
 
-    def update(self, key, paths):
+    def update(self, key: str, entry: CacheableEntry) -> None:
         """
         update cache entry
         """
-        log.verbose("cache update: %s -> %s", key, paths[0])
+        log.verbose("cache update: %s -> %s", key, entry)
         assert key
         self._ensure_load()
-        entries = list(map(path2entry, paths))
-        self.cache[key] = entries
+        self.cache[key] = cache_encode(entry)
         self.save()
 
-    def get(self, key):
+    def get(self, key: str) -> CacheableEntry:
         """
-        get cache entry or None
+        get and validate cache entry or None
         """
         log.verbose("cache query: %s", key)
 
-        def validate(path, checksum):
-            if not path.exists():
-                log.info("removing missing file from cache: %s", path)
-                self.delete(key)
-                return False
-            real_checksum = file_checksum(path)
-            if real_checksum != checksum:
-                log.info("removing invalid cache entry: %s", path)
-                self.delete(key)
-                return False
-            return True
-
-        def entry2path_valid(e):
-            return entry2path(e, validate_fun=validate)
-
         assert key
         self._ensure_load()
-        entries = self.cache.get(key)
-        if not entries:
+        entry = self.cache.get(key)
+        if not entry:
             return None
-        paths = list(map(entry2path_valid, entries))
-        if None in paths:
-            # invalid entry
-            return None
-        return paths
 
-    def delete(self, key):
+        try:
+            entry = cache_decode(entry)
+        except FileMissingError as e:
+            path = e.args[0]
+            log.info("invalidating cache for %s due to missing file: %s",
+                     key, path)
+            self.delete(key)
+            return None
+        except ChecksumError as e:
+            path = e.args[0]
+            log.info("invalidating cache for %s due to failed checksum on: %s",
+                     key, path)
+            self.delete(key)
+            path.unlink()
+            return None
+        return entry
+
+    def delete(self, key: str) -> None:
         """
         delete cache entry
         """
         self.cache.pop(key, None)
         self.save()
 
-    def enabled(self, *targets, cmd=None, use_cache=True):
+    def enabled(self, *targets, cmd: str = None,
+                use_cache: bool = True) -> bool:
         """
         helper to tell and log if caching should be enabled
 
@@ -116,7 +134,7 @@ class ProjectCache:
         optional use_cache utility argument to disable all cache
         """
 
-        def enabled_result(value):
+        def enabled_result(value: bool) -> bool:
             en_str = enabled_str(value)
             if cmd:
                 log.verbose("cache %s for %s", en_str, cmd)
@@ -169,25 +187,45 @@ class ProjectCache:
         return enabled_result(r)
 
 
-def path2entry(path):
+def cache_encode(entry: CacheableEntry) -> CacheEntry:
     """
-    convert a path to corresponding cache entry
-
-    return (fn, checksum) or a list of that on multiple paths
+    Encode entry for use in cache, suported formats are only:
+        - int, bool, str
+        - Path (stored with checksum)
+        - list if all items are a supported format
+        - dict where keys are str, values any supported format
     """
-    return str(path), file_checksum(path)
+    if isinstance(entry, Mapping):
+        result = {}
+        for k, v in entry.items():
+            if k.startswith('__'):
+                raise ValueError(f"Underscore keys are not supported: {k}")
+            result[k] = cache_encode(v)
+        return result
+    elif isinstance(entry, (int, str, bool)):  # str is Iterable
+        return entry
+    elif isinstance(entry, Iterable):
+        return [cache_encode(v) for v in entry]
+    elif isinstance(entry, Path):
+        return {"path": str(entry), "__hash__": file_checksum(entry)}
+    raise NotImplementedError
 
 
-def entry2path(entry, validate_fun=None):
+def cache_decode(entry: CacheEntry) -> CacheableEntry:
     """
-    convert cache entry to a corresponding path
-
-    if validate_fun is specified, it's used confirm file has
-    valid checksum and flush invalid cache entry if it doesn't
+    Decode and validate entry from cache, validating existence and checksum of
+    path entries along the way.
     """
-    fn, checksum = entry
-    p = Path(fn)
-    if validate_fun:
-        if not validate_fun(p, checksum):
-            return None
-    return p
+    if isinstance(entry, dict):
+        if '__hash__' in entry:
+            path = Path(entry['path'])
+            checksum = entry['__hash__']
+            if not path.exists():
+                raise FileMissingError(path)
+            if checksum != file_checksum(path):
+                raise ChecksumError(path)
+            return path
+        return {k: cache_decode(v) for k, v in entry.items()}
+    if isinstance(entry, list):
+        return [cache_decode(v) for v in entry]
+    return entry