+++ apkg/commands/test_dep.py
@@ -13,7 +13,7 @@ log = getLogger(__name__)
 
 
 @cli.command(name="test-dep", aliases=['testdep'])
-@click.argument('input_files', nargs=-1)
+@click.argument('inputs', nargs=-1)
 @click.option('-l', '--list', 'install', default=True, flag_value=False,
               help="list test deps only, don't install")
 @click.option('-u', '--upstream', is_flag=True,
@@ -22,9 +22,8 @@ log = getLogger(__name__)
               help="use test deps from archive")
 @click.option('-d', '--distro',
               help="override target distro  [default: current]")
-@click.option('-F', '--file-list', 'input_file_lists', multiple=True,
-              help=("specify text file listing one input file per line"
-                    ", use '-' to read from stdin"))
+@click.option('-F', '--in-file', 'in_files', multiple=True,
+              help="specify input file, '-' to read from stdin")
 @click.option('--ask/--no-ask', 'interactive',
               default=False, show_default=True,
               help="enable/disable interactive mode")
@@ -42,8 +41,8 @@ def cli_test_dep(*args, **kwargs):
 def test_dep(
         upstream=False,
         archive=False,
-        input_files=None,
-        input_file_lists=None,
+        inputs=None,
+        in_files=None,
         install=True,
         distro=None,
         interactive=False,
@@ -62,9 +61,9 @@ def test_dep(
     distro = adistro.distro_arg(distro, proj)
     log.info("target distro: %s", distro)
 
-    infiles = common.parse_input_files(input_files, input_file_lists)
+    inputs = common.parse_inputs(inputs, in_files)
     archive, _ = parse_archive_args(
-        proj, archive, upstream, infiles)
+        proj, archive, upstream, inputs)
 
     tests = proj.get_tests_for_distro(distro)
     deps = tests.deps